requirements.txt
torch
torchvision
timm
scikit-learn
matplotlib
seaborn
pandas
Pillow
tqdm
numpy
